Update index.html and styles.css to match example
The current design did not match the example. To  practice the assignment
was to create and use the external method with styles.css to change the design
of the div element, to use the internal method to stile the p element and to
use the inline method for the button element.
The final result is corrected to closely match the solution.