Simplifies the stylesheet
In the previous version double style classes were used to
add the shared styles to the main class (button) and define
the specifics (colors) in the secondary classes (black and
yellow). It is simpler to remove the button class and combine
the black and yellow classes in the first decleration,
separating them with a comma.